rename: 멤버 권한 조회 api 폴더 변경
