+++ src/apis/club/getMemberAuth.ts
